[C++] Split util.h to bit_util.h and time_util.h
