+++ src/fury/row/row.h
@@ -21,9 +21,9 @@
 #include "arrow/api.h"
 #include "arrow/status.h"
 #include "fury/row/type.h"
+#include "fury/util/bit_util.h"
 #include "fury/util/buffer.h"
 #include "fury/util/status.h"
-#include "fury/util/util.h"
 
 namespace fury {
 
@@ -165,8 +165,8 @@ public:
   int num_fields() const { return num_fields_; }
 
   bool IsNullAt(int i) const override {
-    return BitUtil::GetBit(buffer_->data() + base_offset_,
-                           static_cast<uint32_t>(i));
+    return util::GetBit(buffer_->data() + base_offset_,
+                        static_cast<uint32_t>(i));
   }
 
   int GetOffset(int i) const override {
@@ -189,11 +189,11 @@ public:
   }
 
   void SetNullAt(int i) override {
-    BitUtil::SetBit(buffer()->data() + base_offset_, i);
+    util::SetBit(buffer()->data() + base_offset_, i);
   }
 
   void SetNotNullAt(int i) override {
-    BitUtil::ClearBit(buffer()->data() + base_offset_, i);
+    util::ClearBit(buffer()->data() + base_offset_, i);
   }
 
   std::string ToString() const override;
@@ -237,8 +237,8 @@ public:
   int num_elements() const { return num_elements_; }
 
   bool IsNullAt(int i) const override {
-    return BitUtil::GetBit(buffer_->data() + base_offset_ + 8,
-                           static_cast<uint32_t>(i));
+    return util::GetBit(buffer_->data() + base_offset_ + 8,
+                        static_cast<uint32_t>(i));
   }
 
   int GetOffset(int i) const override {
@@ -261,13 +261,13 @@ public:
   }
 
   void SetNullAt(int i) override {
-    BitUtil::SetBit(buffer_->data() + base_offset_ + 8, i);
+    util::SetBit(buffer_->data() + base_offset_ + 8, i);
     // we assume the corresponding column was already 0
     // or will be set to 0 later by the caller side
   }
 
   void SetNotNullAt(int i) override {
-    BitUtil::ClearBit(buffer_->data() + base_offset_ + 8, i);
+    util::ClearBit(buffer_->data() + base_offset_ + 8, i);
   }
 
   std::string ToString() const override;
