+++ src/fury/util/BUILD
@@ -4,8 +4,8 @@ load("//bazel:fury.bzl", "COPTS")
 
 cc_library(
     name = "fury_util",
-    srcs = ["status.cc", "buffer.cc", "logging.cc", "util.cc"],
-    hdrs = ["status.h", "buffer.h", "logging.h", "util.h"],
+    srcs = glob(["*.cc"], exclude=["*test.cc"]),
+    hdrs = glob(["*.h"]),
     strip_include_prefix = "/src",
     copts = COPTS,
     alwayslink=True,
@@ -21,8 +21,8 @@ cc_library(
 
 
 cc_test(
-    name = "util_test",
-    srcs = ["util_test.cc"],
+    name = "time_util_test",
+    srcs = ["time_util_test.cc"],
     copts = COPTS,
     deps = [
         ":fury_util",
