+++ src/fury/util/time_util.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <chrono>
+#include <string>
+
+namespace fury {
+
+std::string FormatTimePoint(std::chrono::system_clock::time_point tp);
+
+}
